requirements.txt
chromadb==0.3.21
langchain==0.0.149
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.21.0
